+++ frontend/src/domain/user/model/UserUpdateProfileRequest.ts
@@ -1,0 +1,4 @@
+export default interface UserUpdateProfileRequest {
+    name: string
+    username: string
+}